Update Menu Addition
In this update I changed some menu options as well as moved the seat reset option into the menu, which if selected is prompted AGAIN to make sure this is not a mistake. I created the displayMenu() function (outside of the class) in order to re-display the menu in the seat reset option. If Y is entered, seats are reset, if N, they are not, and are given the menu to enter another choice from the beginning again. if any other char is entered it is invalid and re-prompted